Small Fixes
