+++ e2e/pages/modals/ProtectYourWalletModal.js
@@ -1,11 +1,13 @@
 import TestHelpers from '../../helpers';
 
 const PROTECT_YOUR_WALLET_MODAL_CONTAINER_ID = 'protect-wallet-modal';
-const COLLAPSED_PROTECT_YOUR_WALLET_MODAL_CONTAINER_ID = 'backup-alert';
-const REMIND_ME_LATER_BUTTON_ID = 'notification-remind-later-button';
+import {
+  NOTIFICATION_REMIND_ME_LATER_BUTTON_ID,
+  SECURE_WALLET_BACKUP_ALERT_MODAL,
+} from '../../../wdio/screen-objects/testIDs/Screens/WalletView.testIds';
 export default class ProtectYourWalletModal {
   static async tapRemindMeLaterButton() {
-    await TestHelpers.tap(REMIND_ME_LATER_BUTTON_ID);
+    await TestHelpers.tap(NOTIFICATION_REMIND_ME_LATER_BUTTON_ID);
   }
 
   static async isVisible() {
@@ -17,14 +19,10 @@ export default class ProtectYourWalletModal {
   }
 
   static async isCollapsedBackUpYourWalletModalVisible() {
-    await TestHelpers.checkIfVisible(
-      COLLAPSED_PROTECT_YOUR_WALLET_MODAL_CONTAINER_ID,
-    );
+    await TestHelpers.checkIfVisible(SECURE_WALLET_BACKUP_ALERT_MODAL);
   }
 
   static async isCollapsedBackUpYourWalletModalNotVisible() {
-    await TestHelpers.checkIfNotVisible(
-      COLLAPSED_PROTECT_YOUR_WALLET_MODAL_CONTAINER_ID,
-    );
+    await TestHelpers.checkIfNotVisible(SECURE_WALLET_BACKUP_ALERT_MODAL);
   }
 }
