+++ wdio/screen-objects/WalletMainScreen.js
@@ -18,6 +18,7 @@ import {
   WALLET_ACCOUNT_ICON,
   MAIN_WALLET_VIEW_VIA_TOKENS_ID,
   NOTIFICATION_REMIND_ME_LATER_BUTTON_ID,
+  SECURE_WALLET_BACKUP_ALERT_MODAL
 } from './testIDs/Screens/WalletView.testIds';
 
 import { DRAWER_VIEW_SETTINGS_TEXT_ID } from './testIDs/Screens/DrawerView.testIds';
@@ -76,6 +77,12 @@ class WalletMainScreen {
     );
   }
 
+  get backupAlertModal() {
+    return Selectors.getElementByPlatform(
+      SECURE_WALLET_BACKUP_ALERT_MODAL,
+    );
+  }
+
   async tapSettings() {
     await Gestures.tap(this.drawerSettings);
   }
@@ -126,11 +133,13 @@ class WalletMainScreen {
   }
 
   async tapRemindMeLaterOnNotification() {
-    const timeOut = 3000;
-    await driver.pause(timeOut);
+    await Gestures.tap(this.remindMeLaterNotification, 'MOVETO');
     await Gestures.tap(this.remindMeLaterNotification);
   }
 
+  async backupAlertModalIsVisible() {
+    await expect(this.backupAlertModal).toBeDisplayed();
+  }
   async isVisible() {
     await expect(this.WalletScreenContainer).toBeDisplayed();
   }
